refactor(budgetView): refactor to combine chart and graph components into one
